refactor(simulation): store analysis scalars as floats
Use float storage for energy-deposition and time-of-flight event records, with explicit conversions at the Geant4 event boundary.

+++ include/g4go/simulation/Analysis.hpp
@@ -11,14 +11,14 @@ namespace G4GO::Simulation {
 struct CrystalHitOutput {
     int fEventID{};
     int fModuleID{};
-    double fEnergyDeposit{};
+    float fEnergyDeposit{};
     int fGeneratedPhotonCount{};
 };
 
 struct SensorHitOutput {
     int fEventID{};
     int fSensorID{};
-    double fTimeOfFlight{};
+    float fTimeOfFlight{};
 };
 
 class Analysis final {
